requirements.txt
aiodns==3.0.0
aiofiles==23.1.0
aiohttp==3.8.4
aiomultiprocess==0.9.0
aiosignal==1.3.1
aiosqlite==0.19.0
anyio==3.7.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
blinker==1.6.2
Brlapi==0.8.3
Brotli==1.0.9
capstone==4.0.2
censys==2.2.2
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
colorama==0.4.6
command-not-found==0.3
cryptography==40.0.2
cupshelpers==1.0
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.14
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.2.3
django-barcode==0.1.1
django-filter==23.1
django-materializecss-form==1.1.17
django-post-office==3.6.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.3.0
duplicity==0.8.21
exceptiongroup==1.1.1
fastapi==0.97.0
fasteners==0.14.1
filelock==3.11.0
Flask==2.3.2
frozenlist==1.3.3
future==0.18.2
gyp==0.1
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
importlib-resources==5.12.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonfield==3.1.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
limits==3.5.0
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
multidict==6.0.4
mutagen==1.45.1
mypy-extensions==1.0.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==23.0
paramiko==2.9.3
pathspec==0.11.1
pbr==5.11.1
pexpect==4.8.0
Pillow==9.0.1
pipenv==2023.3.20
platformdirs==3.2.0
plotly==5.14.1
protobuf==3.12.4
psycopg2==2.9.5
ptyprocess==0.7.0
PyAutoGUI==0.9.53
pycairo==1.20.1
pycares==4.3.0
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==1.10.9
pyee==8.2.2
PyGetWindow==0.0.9
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==2.4.7
pyperclip==1.8.2
pyppeteer==1.0.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.28
python-apt==2.4.0+ubuntu1
python-barcode==0.9.0
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python3-xlib==0.15
pytweening==1.0.4
pytz==2023.3
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0
reportlab==3.6.8
requests==2.31.0
retrying==1.3.4
rich==13.4.2
SecretStorage==3.3.1
shodan==1.29.1
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
starlette==0.27.0
stevedore==5.0.0
systemd-python==234
tenacity==8.2.2
tinycss2==1.1.1
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.8.0
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.22.0
uvloop==0.17.0
vboxapi==1.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.6
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.4
wrapt==1.15.0
xdg==5
xkit==0.0.0
XlsxWriter==3.1.2
yarl==1.9.2
yt-dlp==2022.4.8
zenmap==7.93
zipp==1.0.0
